user services cleaned

+++ service/src/main/java/com/epam/impl/UserServiceImpl.java
@@ -10,19 +10,26 @@ import com.epam.UserService;
 
 import java.util.List;
 
+import static com.epam.validator.ServiceValidator.*;
 
-/** validations are needed*/
+
 
 public class UserServiceImpl implements UserService {
 
 
     private static final UserDAO userDAO = DAOFactory.getInstance().createUserDAO();
 
+/**
+ * Implemented methods
+ * */
     @Override
     public boolean registration(User user,String password2) {
-        boolean equals = user.getPassword().equals(password2);
+        if (!validation(user)){
+            return false;
+        }
+        boolean passwordEquals = user.getPassword().equals(password2);
         try{
-            return equals && userDAO.save(user);
+            return passwordEquals && userDAO.save(user);
         }catch (DAOException e){
             throw new ServiceException(e);
         }
@@ -31,6 +38,9 @@ public class UserServiceImpl implements UserService {
 
     @Override
     public UserDTO logination(User user) {
+        if (!validation(user)){
+            return null;
+        }
         try {
             user = userDAO.get(user);
             return (user == null) ? null : new UserDTO(user);
@@ -39,28 +49,83 @@ public class UserServiceImpl implements UserService {
         }
     }
 
+    /**
+     * Methods getter from DB
+     * */
+
+
+    @Override
+    public User get(User user) {
+        if (!validation(user)){
+            return null;
+        }try{
+            return userDAO.get(user);
+        }catch (DAOException e){
+            throw new ServiceException(e);
+        }
+
+    }
+
+
+
     @Override
     public User findUserWithId(long id) {
-        return userDAO.getById(id);
-    }
+        if (!validation(id)) {
+            return null;
+        } try{
+                return userDAO.getById(id);
+            }catch (DAOException e){
+                throw new ServiceException(e);
+            }
+        }
+
 
     @Override
     public boolean updateUser(User user) {
-        return userDAO.update(user);
+        if(!validation(user)){
+            return false;
+        }try{
+            return userDAO.update(user);
+        }catch (DAOException e){
+            throw new ServiceException(e);
+        }
+
     }
 
     @Override
     public boolean deleteUser(User user) {
-        return userDAO.delete(user);
+        if (!validation(user)){
+            return false;
+        }try{
+            return userDAO.delete(user);
+        }catch (DAOException e){
+            throw new ServiceException(e);
+        }
+
     }
 
+    /**
+     * TODO Do I need this?
+     * */
     @Override
     public boolean findUserByLogin(User user) {
-        return userDAO.findUserByLogin(user);
+        if (!validation(user)){
+            return false;
+        }try{
+            return userDAO.findUserByLogin(user);
+        }catch (DAOException e){
+            throw new ServiceException(e);
+        }
+
     }
 
     @Override
     public List<User> getUsers() {
-        return userDAO.getAll();
+        try {
+            return userDAO.getAll();
+        }catch (DAOException e){
+            throw new ServiceException(e);
+        }
+
     }
 }
